+++ plugwise_usb/network/cache.py
@@ -30,7 +30,9 @@ class NetworkRegistrationCache(PlugwiseCache):
             mac: node_type.name for mac, node_type in self._nodetypes.items()
         }
         _LOGGER.debug("Save NodeTypes for %s Nodes", len(cache_data_to_save))
-        await self.write_cache(cache_data_to_save, rewrite=True)  # Make sure the cache-contents is actual
+        await self.write_cache(
+            cache_data_to_save, rewrite=True
+        )  # Make sure the cache-contents is actual
 
     async def clear_cache(self) -> None:
         """Clear current cache."""
@@ -54,7 +56,9 @@ class NetworkRegistrationCache(PlugwiseCache):
                     node_type = None
 
             if node_type is None:
-                _LOGGER.warning("Invalid NodeType in cache for mac %s: %s", mac, node_value)
+                _LOGGER.warning(
+                    "Invalid NodeType in cache for mac %s: %s", mac, node_value
+                )
                 continue
             self._nodetypes[mac] = node_type
             _LOGGER.debug(